refactor: Board 패키지 생성   - Board, BoardCategory Entity 파일 이동
